........
  r253631 | russell | 2010-03-20 14:17:28 -0500 (Sat, 20 Mar 2010) | 2 lines
  
  Resolve a number of FreeBSD build issues.
........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/tilghman-chanvar-backport@253644 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ main/manager.c
@@ -2629,7 +2629,7 @@ int __ast_channel_manager_event(int category, const char *event, int chancount, 
 		now = ast_tvnow();
 		ast_dynamic_str_thread_append(&buf, 0, &manager_event_buf,
 				"Timestamp: %ld.%06lu\r\n",
-				 now.tv_sec, (unsigned long) now.tv_usec);
+				 (long) now.tv_sec, (unsigned long) now.tv_usec);
 	}
 
 	va_start(ap, fmt);
